+++ src/main/java/dev/fulmineo/guild/data/ClientDataManager.java
@@ -4,5 +4,6 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class ClientDataManager {
+	public static Map<String, String> professionsLabels = new HashMap<>();
 	public static Map<String, QuestProfessionRequirement[]> professionRequirements = new HashMap<>();
 }