+++ src/main/java/dev/fulmineo/guild/screen/QuestsScreen.java
@@ -7,6 +7,7 @@ import java.util.Optional;
 import com.mojang.blaze3d.systems.RenderSystem;
 
 import dev.fulmineo.guild.Guild;
+import dev.fulmineo.guild.data.ClientDataManager;
 import dev.fulmineo.guild.data.Quest;
 import dev.fulmineo.guild.data.QuestProfession;
 import dev.fulmineo.guild.screen.QuestsScreenHandler.ProfessionData;
@@ -246,7 +247,9 @@ public class QuestsScreen extends HandledScreen<QuestsScreenHandler> {
 		public void renderTooltip(MatrixStack matrices, int mouseX, int mouseY) {
 			if (this.hovered) {
 				List<Text> tooltip = new ArrayList<>();
-				tooltip.add(this.getQuestProfession().getTranslatedName().formatted(Formatting.GOLD));
+				QuestProfession profession = this.getQuestProfession();
+				String label = ClientDataManager.professionsLabels.get(profession.name);
+				tooltip.add((label != null ? new LiteralText(label) : this.getQuestProfession().getTranslatedName()).formatted(Formatting.GOLD));
 				QuestsScreen.this.renderTooltip(matrices, tooltip, Optional.empty(), mouseX, mouseY);
 			}
 		}